MB-55590: CDC backfill defer snapshot marker until data is available
When the non-history phase of a CDC backfill runs, it may actually
be that no data is available from the start-seqno up to the history
start-seqno. However the backfill code has already pushed a snapshot
marker out. This results in the following sequence of messages for
the client.

* snapshot marker flags=disk
* snapshot marker flags=disk|history
* mutations...

This commit defers the first marker until there is definitely data
resulting in the following two more correct sequences.

* snapshot marker flags=disk
* mutations...
* snapshot marker flags=disk|history
* mutations...

or when only history is available

* snapshot marker flags=disk|history
* mutations...

Note: This fix could be better implemented by not bloating ActiveStream
and allowing the DCPBackfillBySeqnoDisk class to handle this. However
this is difficult to achieve in neo, generally because the snapshot
marker is assumed to appear in the first run of a backfill task. The
master branch would make this much simpler to achieve as backfill has
been changed to bring the marker and data together in 1 run of the
task. When this change enters master it can be revisited so that the
backfill code owns this logic.

Change-Id: Ide76c336f10e6b628a5ac99fd6eb8c3827e49c8e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/186914
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>