[folly] Don't write data from the background threads
Folly don't really like sending data on the socket from
other threads than the one owning the event loop (it could
probably be in the case where it would need to update the
notification status for the socket).

To work around the problem enter the event loop thread and
send the data. This isn't very efficient as we need to
synchronize multiple threads, but it is currently (only?)
used as part of range scans and this makes all unit tests
pass.

We probably want to add some sort of buffering and reuse
the ewb pattern to have the background thread prepare the
data, and send it in the front end thread.

Change-Id: I9dc8abf16c5fbc91e5f2bc20ef391e4fce86aab3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/186785
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>