MB-55557: CDC History scan requires a map of the open collections
With CDC enabled a collection drop doesn't currently have any ability
to fully purge the collection, if the dropped collection spans into
the history window, and the collection has historic data, the older
versions remain in the history window, yet the collection is deemed
dropped (it is no longer in open of dropped collection lists).

When the DCP backfill runs, it transmits these "logically deleted"
items as they are by the Collections::ScanContext logic "alive". This
causes an unknown collection error on the consumer.

The fix for this issue is to enhance the Collections::ScanContext class
so it can use both the set of dropped collections (required) and
optionally be given the set of open collections for deciding what items
are dropped.

The logic becomes:

1) If key@seqno is in the seqno range of dropped collections (and is
in the set of dropped collections) => isLogicallyDeleted is true

Else

2) If the object was created with the list of open collections, the
key@seqno is now checked to see if

2.1) Not in the set of open collections => isLogicallyDeleted is true
2.2) In the set of open collections and seqno < start-seqno
     => isLogicallyDeleted is true

Else the item is deemed alive (in the set and > start-seqno)

Change-Id: I6ccaa104fb30b88bc0adff6b0af96c69ea0f72cc
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/186872
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>