MB-54810: Include overflow samples in mctimings histograms
Make use of the newly-added overflow metrics in timing histograms to
show the number of samples which exceeded the representable range of
the histogram.

Example output from mctimings, where 1000 overflowed samples are
synthesized:

    The following data is collected for GET
    [  0.00 -   7.00]us (0.0000%)    2|
    [  7.00 -   9.00]us (10.0000%)  380| ################
    [  9.00 -  11.00]us (20.0000%)  431| ##################
    [ 11.00 -  12.00]us (30.0000%)  175| #######
    [ 12.00 -  14.00]us (40.0000%)  400| #################
    [ 14.00 -  15.00]us (50.0000%)  340| ##############
    [ 15.00 -  15.00]us (55.0000%)    0|
    [ 15.00 -  16.00]us (60.0000%)  310| #############
    [ 16.00 -  16.00]us (65.0000%)    0|
    [ 16.00 -  17.00]us (70.0000%)  220| #########
    [ 17.00 -  18.00]us (75.0000%)  235| ##########
    [ 18.00 -  18.00]us (77.5000%)    0|
    [ 18.00 -  18.00]us (80.0000%)    0|
    [ 18.00 -  19.00]us (82.5000%)  153| ######
    [ 19.00 -  19.00]us (85.0000%)    0|
    [ 19.00 -  20.00]us (87.5000%)  115| #####
    [ 20.00 -  20.00]us (88.7500%)    0|
    [ 20.00 -  20.00]us (90.0000%)    0|
    [ 20.00 -  21.00]us (91.2500%)   79| ###
    [ 21.00 -  21.00]us (92.5000%)    0|
    [ 21.00 -  22.00]us (93.7500%)   57| ##
    [ 22.00 -  22.00]us (94.3750%)    0|
    [ 22.00 -  23.00]us (95.0000%)   27| #
    [ 23.00 -  23.00]us (95.6250%)    0|
    [ 23.00 -  24.00]us (96.2500%)   23| #
    ...
    [431.00 - 431.00]us (100.0000%)    0|
    [ 60.00 - inf   ]s (overflowed) 1000| ############################################
    Total: 4052 operations

Note the additon of the final 'overflowed' bucket, which reports the
highest trackable value (60s) to Infinity as the range of this bucket.

Also note that the percentiles are calculated based on the buckets
that the histogram, *excluding* any overflow. In other words, "100%"
of the samples based on the information the histogram has correctly
bucketed.

This behaviour might be surprising (why not include overflowed in
percentiles, given we have it now?), but the reason for this is to
maintain consistent percentile iteration - i.e. a given Histogram will
always have buckets at p50, p75, p90, etc. If we re-calculated
percentiles to include the overflow (which the underlying histogram
library which performs the iteration has no knowledge of), then the
reported percentiles would change and this property would be lost.

Change-Id: I9611563a1738a479c15c6d9a7062d99599560566
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/186889
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>