MB-54810: Include overflow samples in cbstats histograms
Make use of the newly-added overflow metrics in timing histograms to
show the number of samples which exceeded the representable range of
the histogram when outputting histograms via cbstats.

Example output from cbstate timings, with 1000 "overflowed" samples
synthesized (each of value 99 minutes):

    $ ./cbstats localhost:12000 -u Administrator -p asdasd runtimes
     WorkLoadMonitor[NonIO] (1045 total)
           0us -    1us : ( 51.1111%)   23
           1us -    1us : ( 51.1111%)    0
    ...
           8us -    8us : ( 97.7778%)    0
           8us -   29us : (100.0000%)    1
          60s - inf     : (overflow ) 1000 ######################################
        Avg             : (93m:55s)

Note the addition of the final 'overflowed' bucket, which reports the
highest trackable value (60s) to Infinity as the range of this bucket.

Also note that the percentiles are calculated based on the buckets
that the histogram, *excluding* any overflow. In other words, "100%"
of the samples based on the information the histogram has correctly
bucketed.

This behaviour might be surprising (why not include overflowed in
percentiles, given we have it now?), but the reason for this is to
maintain consistent percentile iteration - i.e. a given Histogram
should have buckets at p50, p75, p90, etc. If we re-calculated
percentiles to include the overflow (which the underlying histogram
library which performs the iteration has no knowledge of), then the
reported percentiles would change and this property would be lost.

Finally, note that the Avg (mean) field is calculated to include the
overflow samples.

Change-Id: I827d82aa7272d2b464e3993cea514dce5363233a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/186957
Well-Formed: Restriction Checker
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>