MB-54810: Include overflow samples in Prometheus histograms
Make use of the newly-added overflow metrics in timing histograms to
show the number of samples which exceeded the representable range of
the histogram, and their sum.

Example output from _prometheusMetricsHigh endpoint, with 1000
"overflowed" samples synthesized (each with a duration of 999
seconds):

    http --auth Administrator:asdasd http://localhost:11996/_prometheusMetricsHigh|rg kv_cmd_duration_seconds | rg -w GET
    kv_cmd_duration_seconds_count{bucket="default",opcode="GET"} 1005 1676996155000
    kv_cmd_duration_seconds_sum{bucket="default",opcode="GET"} 999000.001152 1676996155000
    kv_cmd_duration_seconds_bucket{bucket="default",opcode="GET",le="0.000001"} 0 1676996155000
    kv_cmd_duration_seconds_bucket{bucket="default",opcode="GET",le="0.000002"} 0 1676996155000
    ...
    kv_cmd_duration_seconds_bucket{bucket="default",opcode="GET",le="65.011712"} 5 1676996155000
    kv_cmd_duration_seconds_bucket{bucket="default",opcode="GET",le="+Inf"} 1005 1676996155000

Note that up to the highest bucket the histogram tracks (65s) there
are cumulatively 5 samples; however at the "infinity" bucket there
are an additional 1000 samples. These are also reflected in the _count
and _sum metrics - i.e. Prometheus calculating the mean as _sum /
_count will include the overflowed samples.

Change-Id: I3491997ce50de756f0d414ddf345bddf8e07fd97
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/186947
Well-Formed: Restriction Checker
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>