MB-55600: MagmaKVStore::getRange fails for system events+KEYS_ONLY
The following fix:
https://review.couchbase.org/c/kv_engine/+/150166

Made a change to MagmaKVStore::makeItem so that if the key is a system
key always create an Item with a value. This function is shared with
many MagmaKVStore functions, in particular getRange

A KEYS_ONLY getRange that encounters a system key will fail (detected
by Nexus backend) as it creates a value on Magma, but not on couchstore.
The value itself is empty because the getRange KEYS_ONLY usage of
makeItem passes an empty value slice, which this fix checks for to skip
the value_t creation.

Change-Id: Ia7ff5e650f7781fc566210dd31fd25af24a0f8d2
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/186911
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>