MB-55507: Checkpoint can't store duplicate of expelled item
There is an invariant in KV: a Checkpoint can always store only one
single revision for some key.

For enforcing that invariant, a Checkpoint:

  1. Just queues a mutation if that key isn't in the key-index
  2. Deduplicates the existing mutation if the key is in the key-index
     (ie, some revision for that key is already queued in the
      checkpoint)

Obvisouly (2) doesn't happen for collection(history:true) items. A new
checkpoint is created and the new (non-dedup) mutation is queued into
the new checkpoint.

But, we have a special path in the Checkpoint's deduplication code that
handles the case where the Checkpoint contained a mutation for some key,
but that mutation has been expelled (for memory recovery).
In that case the original (pre-CDC) behaviour is to just queue the new
mutation in the checkpoint. Ie, the old one is not physically in the
checkpoint anymore, so the Checkpoint/single-revision invariant still
holds.

The latter doesn't work for CDC. Queueing a new revision of a key into
the same checkpoint as the previous revision means that DCP will stream
the two revisions as being part of the same physical checkpoint. Which
will fail replica as it will try to queue the duplicates (which are not
deduplicable) into the same checkpoint.

Change-Id: I6e9e0b27239ccaeb3e0f4c14a1806de189949775
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/186641
Well-Formed: Restriction Checker
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>