MB-55678: StreamRequest flag STRICT_VBUUID should be considered valid
After the change to validate DCP StreamRequest flags (f5a17f60f),
clients attempting to use flag STRICT_VBUUID (0x20) begin to be
rejected. This is a due to a latest issue in when that flag was aded -
it was omitted from the valid flags for AddStream command (9bd5a5796).

This had no immediate effect, as the AddStream command (as generated
by ns_server) doesn't use that flag. However, as part of MB-55441 we
re-used the same validation code for DCP StreamRequest, and hence that
flag was incorrectly rejected.

Fix by including in valid flags; also enhance the unit test to check
_all_ possible flags.

Change-Id: Ifab968bbe9ee85b37b4d501cecbc71b28458f79c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/187285
Well-Formed: Restriction Checker
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>