MB-55694: Update replicaHTMemory stat when clearing HashTable
We currently do not update the replicaHTMemory stat when we destroy an
ephemeral vBucket and that can result in us failing to track memory
usage correctly and failing to run the item pager when we should.

This was spotted in the rollback case which deletes vBuckets if we
rollback to zero (we always roll back to zero for ephemeral vBuckets).

Change-Id: I6941e4fb6d6681f2c81a806c5ed5d6537315d08e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/187361
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker