MB-55584: Ensure HashTable lock has not been released in more methods
Adds additional sanity checks in HT methods.

A crash has been observed (linked MB) which suggests HT state is not
being managed correctly. A StoredValue, found while walking the HT, is
later found to be _absent_ from the HT even though the appropriate
lock has been held for the duration.

The root cause has not been identified, but to tighten checks around
hashtable locks, add checks to more methods which verify the provided
lock has not been released.

In the end, a released HT lock may not be at fault; this is largely to
try catch such a situation sooner to aid debugging if it _does_ occur,
now or in the future.

Change-Id: I2887ba444c142fec5f9a4fc7b64577152cf2b01a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/187437
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>