MB-55584: Tighten contract of HashTable methods taking StoredValue*
If called with an null chain ptr, hashChainRemoveFirst would
erroneously dereference the ptr to check getNext().

All use of hashChainRemoveFirst is from actors which are holding a hash
bucket lock, and are attempting to remove a StoredValue _known_ to be
present in the given chain. Thus, the unconditional dereference was not
previously encountered - the chain can't logically be empty if it
contains the requested SV.

This hidden precondition was not enforced; indeed hashChainRemoveFirst
was prepared to return a nullptr if no matching value was found,
allowing the caller to handle this logic error.

An unknown issue (see MB) led to a violation of this precondition -
attempting to remove a SV from an "empty" chain (nullptr).

While this does not address the _cause_ of this issue,
hashChainRemoveFirst can more explicitly enforce the above contract;
the method may only be called if the requested value is known to exist
in the chain.

As it is now invalid to ever call hashChainRemoveFirst and _not_ find
the given value, and the existing callers only ever do a ptr comparison
on the value, the interface can be further simplified into taking a
StoredValue reference rather than a predicate.

This can then be propagated back up the stack; unlocked_del and
unlocked_ejectItem both call hashChainRemoveFirst, and in turn both
_also_ require a non-null StoredValue ptr.

Change-Id: I54975400fce1dc638f3e228f313846bb704bf471
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/187438
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>