MB-55672: Convert RangeScan to perform network IO on frontend
Before this commit the "send" of the data read by a RangeScan was
occurring on the IO task. With this commit now all data read by
a RangeScan is "sent" by the frontend executor.

RangeScan's are converted to follow the established notify IO complete
pattern.

Front-end thread receives a client's range-scan-continue request
and execution enters ep-engine continueRangeScan passing "success" as
the currentStatus.

ep-engine now schedules an IO task to move the scan forwards and
returns would_block. Front-end thread is now free to do other work.

IO thread will run the scan until some event triggers the task to
yield. E.g. a limit was reached or the responseBuffer which stores
the scanned data is at or over the configured limit. What ever the
event was, this is returned to the front-end thread via
notifyIOComplete.

Front-end thread re-processes the range-scan-continue and execution
enters ep-engine passing the IO complete status as currentStatus.

ep-engine detects the IO complete phase by reading back an object
which was stashed in the engine-specific then if currentStatus is:

success: A continue is still running but the responseBuffer needs
sending. The responseBuffer is "drained" to the client via
Cookie::sendResponse and an IO task rescheduled - ep-engine
returns would_block. The RangeScan remains in the Continue state.

range_scan_more: A continue reached a requested limit or was throttled.
The responseBuffer is "drained" to the client via
Cookie::sendResponse and the request is complete. The RangeScan is
changed to be Idle.

range_scan_complete: A continue reached the scan end. The
responseBuffer is "drained" to the client via Cookie::sendResponse and
the request is complete. The RangeScan is changed to be Complete and
removed from the map of available RangeScan objects.

other status: A continue encountered an error. The responseBuffer is
cleared (not sent). The RangeScan is changed to be Cancelled and
removed from the map of available RangeScan objects.

With this change some of the RangeScan state management is simplified
in that the RangeScanOwner does all of the state changes in response
to the status of a continue - no longer does the IO task drive the
state to be Cancelled/Complete.

With this change the performance of a RangeScan (as measured by the
kv_engine/rangescan tool) is reduced. Primarily as between each 8K
of keys/values there is a "cycle" of io-complete and task scheduling.
Prior to this change, the IO task would keep on running with no
reschedule required.

Data from the rangescan tool suggest that with this commit there is a
~7% reduction in rangescan throughput.

Test used a single node with 16 vbuckets, 420MB of keys will be
returned by the scan (total of all vbuckets). The scan used no limits,
i.e. a single continue request progresses the scan from start to end.

Time to perform the scan (no limits), average of 5 runs.

pre   | post
------|------
37.8s | 40.6s

This difference can be removed by increasing the RangeScan buffer size
and will be done in a later commit.

Change-Id: I17566daa37700a4caaf1f85152a4185929b29cf0
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/187490
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>