MB-55391: Don't leave Cursors in CM if not moved to ActiveStream
In ActiveStream::scheduleBackfill_UNLOCKED an exception was thrown
before the result of registerCursor was assigned to
ActiveStream::cursor. This resulted in the cursor being orphaned - it
existed in CheckpointManager but was not associated with any Stream,
and hence was not advanced (via getItemsForCursor) - but also was not
subject to cursor-dropping so resulted in the CheckpointManager
getting stuck at the checkpoint high watermark.

Address this problem (and similar issues) by having CursorRegResult
own the Cursor object, until the user takes ownership of it (via new
takeCursor() method). If CursorRegResult is destroyed and still owns
the cursor; then remove it from CheckpointManager.

Note: An alternate approach would be to implement the "auto remove"
      logic to CheckpointCursor itself (in its dtor); then say have
      CheckpointManager only hold a weak_ptr to CheckpointCursor
      (currently holds a shared_ptr). That would solve the problem at
      a higher level and for all uses of Cursor / CheckpointCursor,
      even after cursor has been moved from CursorRegResult. For
      example right now we rely on ActiveStream's dtor "remembering"
      to remove the cursor from CheckpointManager.

      While such an approach certainly has merit, I feel it's more
      invasive as we would be changing the ownership model of
      CheckpointManager / CheckpointCursor, so at this stage in Elixir
      development, the current approach - while less "complete" is
      smaller in scope and at lower risk of introducing any
      regressions.

Change-Id: If292d10d03cbddeb34cd567e7ce5110a5d58db70
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/187503
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>