MB-55815: Expose per-bucket RSS stats
Per-bucket RSS is used to calculate the fragmentation size for the
bucket and to inform DefragmenterTask scheduling for the
auto_pid/auto_linear modes.

As such, it is useful to have the historical values of that stat in
Prometheus (we already have it in cbstats memory), both as additional
context for memory fragmentation issues, and for future tuning of the
DefragmenterTask itself.

Change-Id: Iae8bd73764ea038fe1564cb23ba29447fdf4d29c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/187724
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>