MB-55837: MagmaKVStore preparePendingRequests must compare DiskDocKey
The sorting is comparing the Item's StoredDocKey, which means that
prepares are placed incorrectly in the output. Magma then sees input
which is incorrect as the prepares (as DiskDocKey) are now in the
incorrect place. This leads to the incorrect number of update
callbacks to KV and an incorrect item count.

Change-Id: I9f587c4447e8b00f8be0bce57fc8782d391affd6
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/187781
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>