MB-53898: Make expiries from compaction fetch values immediately
Previously, if compaction attempted to expire a non-resident value, it
would queue a bgfetch to bring the current version of the value into
memory.

This lead to contention with bgfetches from the frontend; compaction
could generate a large number of bgfetches, which would delay serving
fetches for frontend ops.

As compaction is already performed on background threads, queuing a
bgfetch for a different task to complete is not strictly necessary.

Make compaction expiries read values from disk "inline" one at a time,
avoiding queuing any bgfetches.

This adds config param

 compaction_expiry_fetch_inline

which defaults to true. This can be set to false (dynamically) to
return to the previous behaviour.

Change-Id: I2e85a78459828dad8cb1963cb93cfed9c4e45652
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/189158
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker