MB-53898: Add ep_bg_fetched_compaction metric
Previously, bgfetches triggered by compaction (to fetch the current
version of a document for expiry) were not tracked such that they could
be conveniently distinguished from bgfetches for frontend ops.

To resolve this, add a new monotonic counter which is incremented for
every document compaction attempts to fetch from disk (regardless of if
it succeeds).

Change-Id: I56e960eda39787c034c7eaee4fd439a3f1349bd7
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/189159
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Well-Formed: Restriction Checker