Add support for lambda config change listeners
Allows users interested in a config param changing to define listeners
as a lambda.

Previously, a subclass of ValueChangedListener was required. As a
result of this verbosity, a single listener type is often used to
handle changes in several config keys, even though an instance per key
must be constructed anyway (managed by unique_ptrs).

For greater convenience, add `addValueChangedCallback` taking a
templated Callable. This permits lambdas to be used.

Note, as the provided lambda is currently stored in an std::function,
generic lambdas cannot be used. As a given lambda is likely only
associated with a single key (and thus needs only handle a single type)
this is acceptable for now.

Change-Id: I4f8b3d2acecc9cae95e17efa6054922c2cc51a78
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/188036
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>