MB-55894: Correct the logging of CanDeduplicate
This issue has only been observed in server builds, but the logging
of CanDeduplicate appears "corrupt". I suspect that the correct
ostream operator is not being found and the value is logged as a char.

Rather than trying to figure this out on the server build, just call
to_string which should resolve correctly to to_string(CanDeduplicate)

Change-Id: Ie253a22553b57db337ab68194d0b2200e1ce07ab
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/187960
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>