MB-55943: Expose the defragmenter task sleep time as a stat
We already have some visibility on what the defragmenter is doing, by
means of the _num_visited stats. Those are not a great proxy for how
frequently the defragmenter runs however.

The visitation rate will depend on what the defragmenter is doing. If
we are visiting _and_ moving items, the visitation rate will drop,
because moving items is more expensive than just visiting.

Expose the raw task sleep times as stats, as these tell us what the
auto defragmenter modes are doing.

Change-Id: I5e465ab2ee1fdf8a2f4b7185091c8841304d72a3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/188103
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>