MB-55943: [BP] Expose the defragmenter task sleep time as a stat
Backport change to 7.1.4-MP2.

We already have some visibility on what the defragmenter is doing, by
means of the _num_visited stats. Those are not a great proxy for how
frequently the defragmenter runs however.

The visitation rate will depend on what the defragmenter is doing. If
we are visiting _and_ moving items, the visitation rate will drop,
because moving items is more expensive than just visiting.

Expose the raw task sleep times as stats, as these tell us what the
auto defragmenter modes are doing.

Change-Id: Ic1879eba01da74126296a4213cf1c4e1e14d84c9
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/189256
Well-Formed: Restriction Checker
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>