MB-55943: [BP] Reduce the defragmentation age thresholds to 1
Backport change to 7.1.4-MP2.

The StoredValue and Blob age fields are incremented whenever the
DefragmenterTask visits a value in the HT. The original intention was to
protect items which have just been added into the HT for some period of
time, relying on the assumption that older allocations are more likely
to contribute to fragmentation.

However, we've seen cases where this is not necessarily true and the
current thresholds of 10 are likely too high. Also, we don't really
want to have to visit each item 10 times before we can defragment it.

Reduce the threshold to 1. This gives each item one defragmenter pass of
"protection" before we can reallocate it. The reason we haven't
completely removed the threshold is that in bulk workload scenarios, the
newly inserted items will not be fragmented, and we don't want to waste
CPU trying to defragment them the first time round, as defragmenting is
much slower that visiting.

Change-Id: Ie2b81aac3bf87ba8aff2672f3d2719f4315ce80d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/189257
Well-Formed: Restriction Checker
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>