MB-55943: Set the auto_pid proportional factor to a higher value
Set the proportional factor Kp (defragmenter_auto_pid_p) to 52.

This value was calculated based on auto thresholds and sleep times, such
that we reduce sleep time to min_sleep (0.6s) when we reach the
higher_threshold.

Kp * (upper - lower) = max_sleep - min_sleep, solving for Kp
    Kp = (10 - 0.6) / (0.25 - 0.07) = 52

We're keeping the integral term the same, so the only change in
behaviour is we reduce sleep to min_sleep when scoredFragmentation is
0.25. We'd still end up reducing the sleep to min_sleep in cases where
scoredFragmentation is just above the threshold, as before.

Change-Id: I3a575c4f3a8bdc0bc2c8eeeb315beda3063cbb7d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/188106
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>