MB-55943: [BP] Set the auto_pid proportional factor to a higher value
Backport change to 7.1.4-MP2.

Set the proportional factor Kp (defragmenter_auto_pid_p) to 52.

This value was calculated based on auto thresholds and sleep times, such
that we reduce sleep time to min_sleep (0.6s) when we reach the
higher_threshold.

Kp * (upper - lower) = max_sleep - min_sleep, solving for Kp
    Kp = (10 - 0.6) / (0.25 - 0.07) = 52

We're keeping the integral term the same, so the only change in
behaviour is we reduce sleep to min_sleep when scoredFragmentation is
0.25. We'd still end up reducing the sleep to min_sleep in cases where
scoredFragmentation is just above the threshold, as before.

Change-Id: Iaa16d19f872cce5c2abf6623a73d9ab13e3b7662
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/189258
Well-Formed: Restriction Checker
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>