MB-55919: Set the correct Item::deduplicate flag for queue_op::abort
The Item::deduplicate flag must be set for historical:true collections
items if history is enabled at bucket level.

We missed to do that for Abort operations. The result is that Abort
might be deduplicated at flusher and never persisted to disk. Eg, if CM
stores a sequence for the same key like:
 - 1:pre
 - 2:abr
 - 3:pre
 - 4:cmt
then 2:abr is deduplicated and disk gets persisted only 1:pre, 3:pre,
4:cmt.

The consequence is that the historical disk snapshot generated on DCP
misses 2:abr and sends 1:pre, 3:pre in sequence. Which fails sanity
checks at replica.

Change-Id: Ia571c020b83ce6fb4abaee0b4a94a65ed591b2fe
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/188146
Well-Formed: Restriction Checker
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>