MB-55943: Add test for the new default auto_pid configuration
The updated value for the proportional terms was chosen to be the value
which given minSleep for fragmentation at upper_threshold and maxSleep
for fragmentation at lower_threshold when the integral term is 0.

This essentially makes the auto_pid behaviour the same as that of the
auto_linear mode in the initial state of the PID controller.

Add a test to make sure we're always in line with that, to guard against
inconsistent changes to the parameters this value was computed from.

Change-Id: Ib92e1f064e474c3596e287ebd96d8b2a6bff5046
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/188147
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>