MB-55939: Catch any exceptions thrown in MutationLog dtor
If the MutationLog fails to write to disk when it is being destroyed
(while writing out the header and closing the file), then it throws a
WriteException. This is not caught by the calling ItemAccessVisitor
ctor (quite reasonably as destructors should never fail) and results
in process termination as the uncaught exception is propogated up the
stack.

Address this by catching any thrown exceptions inside MutationLog dtor
and forcefully closing the file - there's not much else we can
sensibly do at this point.

Change-Id: Ifc7fad6c53df9f202596dffc00a0ee122a16e3ca
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/188159
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>