Ensure config listener callbacks handle the correct type
Config params will always have a value of a particular type.

The older ValueChangedListener style listeners with overridden virtual
methods couldn't easily enforce that the supplied listener actually
handled the config param's type.

Furthermore, as they are quite verbose, listener types were often
reused to listen on several config keys. This made it even easier to
mis-handle a config key. E.g., the size_t method could be overridden,
and check for a config key, but the config value could actually be
ssize_t (which might _also_ be overridden, but not check for the
correct key).

For lambda/std::function callbacks this is much more straightforward;
the callback handles a single type and listens for changes in a single
key.

Enforce that the provided callback can handle the type of the
associated config param.

Change-Id: I33588222aed1cb61152b7972579221d4ef4ca357
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/188169
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>