MB-55958: Item's with no value should have no datatype (raw 0)
When creating an item with no value do not copy the datatype, this
covers the broadest amount of code paths rather than a targetted fix
just for StoredValue::toItemAbort (as per the MB). Code review confirms
that other usage of toItemInfo with the NoValue option does not expect
the datatype, including DCP's "key only + underlying datatype" mode.

For context, the "key only + underlying datatype" does not request
Item's with no value from the hash-table, this can be checked by
looking at the usage of getInternal from CacheCallback and comments
also in ActiveStream::isKeyOnly.

Additionally MagmaKVStore is changed to only check for valid items
after the Item operation type is updated to abort/prepare. This was
some source of confusion in the MB where the problematic abort was
logged as a normal looking delete.

Note: This commit was originally intended to be submitted with the
following change in CheckpointManager::queueDirty

 #if CB_DEVELOPMENT_ASSERTS
    if (qi->isDeleted() && qi->getNBytes() == 0 &&
        qi->getDataType() != PROTOCOL_BINARY_RAW_BYTES) {
        throw std::invalid_argument(fmt::format(
                ".some error. {}", *qi));
    }
 #endif

Which lead to the code in kv_bucket_test.cc to ensure we don't use the
low level API's to violate that assert. However the test function
testSanitizeOnDiskDeletedDocWithIncorrectXATTR requires to violate that
condition, so for now the exception is removed.

Change-Id: Id80fd802572bd297a42c10149167a1b2b2b28bc5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/188232
Well-Formed: Restriction Checker <build-team@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>