MB-56180: Reimplement throttle algorithm
The specification for throttling changed in the way that we
shouldn't limit a tenant to a certain limit iff there is available
capacity on the node.

This patch introduce 3 different tunables:

    node capacity    - This is the max units per second available
                       on the node and all of the buckets eat off
                       this limit

    reserved         - This is a reserved minimum on a per bucket
                       base that the bucket should have. No other
                       tenant may consume resources allocated for
                       this bucket

    hard_limit       - This is a maximum number of units the bucket
                       may have. This value is optional, and if not
                       set one may use up to all available resources.

Every second we subract all "reserved" units from each tenant
from the node capacity, and this is the "free pool" where all
tenants may allocate from once they exceed their soft limit.
When the free pool is empty everything gets throttled (unless
the connection holds the Unthrottled privilege)

There is currently an unused "read throttling factor" (rtf) and
"write throttling factor" (wtf) which is set to 1.0 for each
command. The thought was that one may want to reduce rtf for
memory resident items to reflect that they're cheaper to operate
on, and for instance increase wtf for durable ops (currently
they're hardcoded to 2x for a different reason (metering cost))

Change-Id: I13596eec172b56bca743815e5f298c55131bc562
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/188500
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>