MB-56081: Create McProgramGetopt
All of the mc programs use the same subset of options to configure
a connection to the server. To avoid having to duplicate this logic
to all of the tools this patch introduce a common "class" they
may use to parse command line arguments (and look up hostname,
get password from console or environment variable etc).

The class is based upon getopt and getopt_long instead of flipping
to boost::program_options due to:

  1) boost::program_options added a lot of extra "boilerplate"
     code needed to set up the options, parse the options and
     finally look for the program arguemts.

  2) It was non-intuitive to get boost::program_option to
     correctly deal with options which took an optional
     argument. Our option --tls may be used like:
     --tls           - Just use a secure channel
     --tls=cert,key  - Use secure channel with the provided
                       certificate and key.

     From my attempts to play around with "default_value()"
     and "implicit_value()" I could get it to not pass
     the next argument as the value for --tls, but it would
     always mark it as "set". And all of this complexity
     would have to be set for any option working like this,
     causing the boilerplate code in all of the applications
     to grow.

As part of flipping to the unified command line parsing all of
the old deprecated command line options was removed

Change-Id: If6681d09aab1534af3f7ffadbffa145f70142aff
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/188538
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>