[BP] MB-56084: Bump the lastReadSeqno when client is not collection-aware
Backport change to 7.1.5.

A legacy stream with an in-memory phase can fail to backfill if the
purgeSeqno > _default.highSeqno. This could happen even if startSeqno=0,
as the test shows.

If this situation occurred with a collection-aware client, we'd send a
SeqnoAdvanced and advance the lastReadSeqno to the maxScanSeqno that the
disk backfill saw.

However, if the client is not collection-aware, we currently leave the
lastReadSeqno = snapEndSeqno of the last snapshot, which may be below
the purgeSeqno.

Then, we schedule another backfill (normally to catch up with mutations
up to the cursor seqno), but that second backfill starts at the
snapEndSeqno (== lastReadSeqno) + 1. If that is less than or equal to
the purgeSeqno, the second backfill fails with reason=Rollback, and
the stream fails.

Change-Id: I1c171c2695543b93b2a9e9781a6f0b69f2925bc9
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/192379
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/193089
Reviewed-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>