MB-56067: Move bucket destruction to a YieldingTask
Bucket destruction runs in a NonIO task, but may block waiting for
connections to go away.

If enough buckets are deleted at the same time, _all_ NonIO threads may
be blocked by bucket deletion tasks.

Bucket deletion may be blocking waiting for DCP connections to
disconnected, but for that to happen the ConnManager task needs to run
to destroy the ConnHandler.

Unfortunately, that is also a NonIO task, so this may deadlock.

To avoid this situation, make the bucket deletion task snooze while
waiting for connections to die, so that other tasks may run.

Change-Id: Id0081249e08c900063e20609527df70efce3c141
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/188727
Well-Formed: Restriction Checker
Tested-by: James H <james.harrison@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>