MB-55672: RangeScan reserve the frontend/IO task responseBuffer
Profiling data of a key-scan showed that malloc and memcpy were
high in the profile of the range-scan IO threads, with the majority
tracing back to the push_back onto the responseBuffer.

To reduce these we can just reserve the vector in handleKey/handleItem
so that the vector is at the desired config size.

Change-Id: I035986eb67565260b0b45d18d251bb8617169f06
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/188816
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>